requirements.txt
Django==1.11
django-crispy-forms==1.6
django-phonenumber-field==0.7.2
phonenumbers==7.1.0
